apps/api/requirements.txt
fastapi==0.109.0
uvicorn[standard]==0.27.0
python-multipart==0.0.6
pydantic==2.5.3
pydantic-settings==2.1.0
python-pptx==0.6.23
Pillow==10.2.0
sqlalchemy==2.0.25
asyncpg==0.29.0
psycopg2-binary==2.9.9
alembic==1.13.1
celery==5.3.6
redis==5.0.1
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
bcrypt==4.1.2
# S3
boto3==1.34.14
aioboto3==12.3.0
# Observability
python-json-logger==2.0.7
asgi-correlation-id==4.3.0
